requirements.txt
numpy
pandas
scikit-learn
seaborn
python-dotenv
streamlit
google-generativeai
webvtt-py